switch to modlookup for fetching channel lists
Twitch only gives us 20 followed channels at a time while some people
follow 100s of channels. Checking whether a user is a mod in each
channel is a tedious process (ignoring the multiple requests to get the
entire follow list).

Luckily a service exists that can tell us who our user is a mod for and
it should provide us with enough results to show all channels in one go.
This dramatically improves the user experience.